+++ src/engine/ecs/systems/sdl_event_dispatch_system.cpp
@@ -14,7 +14,7 @@ namespace chestnut
 
     }
 
-    void CSDLEventDispatchSystem::update( uint32_t deltaTime ) 
+    void CSDLEventDispatchSystem::update( float deltaTime ) 
     {
         SDL_Event sdlEvent;
