+++ src/engine/ecs/systems/timer_system.hpp
@@ -14,7 +14,7 @@ namespace chestnut
     public:
         CTimerSystem( CEngine& engine );
 
-        void update( uint32_t deltaTime ) override;
+        void update( float deltaTime ) override;
     };
 
 } // namespace chestnut
